requirements.txt
langchain
langchain-core
langchain-ollama
langchain-community
langchain-mcp-adapters
python-dotenv
fastmcp